[ModernSearch] Added a 'Container Tree' refiner template.
