+++ search-extensibility-library/src/libraries/CustomComponent.tsx
@@ -81,8 +81,10 @@ export class MyCustomComponentWebComponent extends BaseWebComponent {
     }
  
     public async connectedCallback() {
- 
+        
        let props = this.resolveAttributes();
+
+       // You can use this._ctx here to access current Web Part context
        const customComponent = <CustomComponent {...props}/>;
        ReactDOM.render(customComponent, this);
     }
