+++ search-parts/config/package-solution.json
@@ -5,7 +5,7 @@
         "id": "890affef-33e0-4d72-bd72-36399e02143b",
         "version": "3.9.0.0",
         "includeClientSideAssets": true,
-        "skipFeatureDeployment": false,
+        "skipFeatureDeployment": true,
         "isDomainIsolated": false
     },
     "paths": {
