+++ search-parts/src/models/RefinerTemplateOption.ts
@@ -7,7 +7,7 @@ enum RefinerTemplateOption {
     Persona = 6,
     FileType = 7,
     FileTypeMulti = 8,
-    Folders = 9
+    ContainerTree = 9
 }
 
 export default RefinerTemplateOption;
