+++ search-parts/src/webparts/searchRefiners/SearchRefinersWebPart.ts
@@ -66,7 +66,6 @@ export default class SearchRefinersWebPart extends BaseClientSideWebPart<ISearch
     let renderElement = null;
     let availableRefiners = [];
     let queryKeywords = '';
-    let queryModification = '';
     let selectedProperties: string[] = [];
     let queryTemplate: string = '';
     let resultSourceId: string = '';
@@ -84,7 +83,6 @@ export default class SearchRefinersWebPart extends BaseClientSideWebPart<ISearch
           selectedProperties = (searchServiceConfig.selectedProperties) ? searchServiceConfig.selectedProperties : [];
           queryTemplate = (searchServiceConfig.queryTemplate) ? searchServiceConfig.queryTemplate : '';
           resultSourceId = searchServiceConfig.resultSourceId;
-          queryModification = searchResultSourceData.queryModification;
         }
       }
 
@@ -104,8 +102,7 @@ export default class SearchRefinersWebPart extends BaseClientSideWebPart<ISearch
           language: this.context.pageContext.cultureInfo.currentUICultureName,
           query: queryKeywords + queryTemplate + selectedProperties + resultSourceId,
           themeVariant: this._themeVariant,
-          userService: this._userService,
-          queryModification: queryModification
+          userService: this._userService
         } as ISearchRefinersContainerProps
       );
     } else {
@@ -265,8 +262,8 @@ export default class SearchRefinersWebPart extends BaseClientSideWebPart<ISearch
                 text: strings.Refiners.Templates.FileTypeMutliValueRefinementItemTemplateLabel
               },
               {
-                key: RefinerTemplateOption.Folders,
-                text: "Folders"
+                key: RefinerTemplateOption.ContainerTree,
+                text: strings.Refiners.Templates.ContainerTreeRefinementItemTemplateLabel
               }
             ]
           },
