+++ search-parts/src/webparts/searchRefiners/components/Layouts/IFilterLayoutProps.ts
@@ -55,11 +55,6 @@ interface IFilterLayoutProps {
    * The current theme variant
    */
   themeVariant: IReadonlyTheme | undefined;
-
-  /**
-   * The current query modification
-   */
-  queryModification: string;
 }
 
 export default IFilterLayoutProps;
