+++ search-parts/src/webparts/searchRefiners/components/Layouts/Vertical/Vertical.tsx
@@ -190,7 +190,6 @@ export default class Vertical extends React.Component<IFilterLayoutProps, IVerti
           themeVariant={props.themeVariant}
           selectedValues={selectedFilterValues}
           userService={this.props.userService}
-          queryModification={this.props.queryModification}
         />
       );
     });
