+++ search-parts/src/webparts/searchRefiners/components/Templates/TemplateRenderer.tsx
@@ -5,7 +5,7 @@ import DateRangeTemplate from "./DateRange/DateRangeTemplate";
 import FixedDateRangeTemplate from "./FixedDateRange/FixedDateRangeTemplate";
 import PersonaTemplate from "./Persona/PersonaTemplate";
 import FileTypeTemplate from "./FileType/FileTypeTemplate";
-import FoldersTemplate from "./Folders/FoldersTemplate";
+import ContainerTreeTemplate from "./ContainerTree/ContainerTreeTemplate";
 import { IRefinementResult, IRefinementValue } from "../../../../models/ISearchResult";
 import RefinementFilterOperationCallback from '../../../../models/RefinementValueOperationCallback';
 import IUserService from '../../../../services/UserService/IUserService';
@@ -58,11 +58,6 @@ export interface ITemplateRendererProps {
    * The current theme variant
    */
   themeVariant: IReadonlyTheme | undefined;
-
-  /**
-   * The current query modification
-   */
-  queryModification: string;
 }
 
 export default class TemplateRenderer extends React.Component<ITemplateRendererProps> {
@@ -159,9 +154,8 @@ export default class TemplateRenderer extends React.Component<ITemplateRendererP
         />;
         break;
 
-      case RefinerTemplateOption.Folders:
-          renderTemplate = <FoldersTemplate
-            queryModification={this.props.queryModification}
+      case RefinerTemplateOption.ContainerTree:
+          renderTemplate = <ContainerTreeTemplate
             refinementResult={this.props.refinementResult}
             onFilterValuesUpdated={this.props.onFilterValuesUpdated}
             shouldResetFilters={this.props.shouldResetFilters}
