+++ search-parts/src/webparts/searchResults/SearchResultsWebPart.ts
@@ -338,7 +338,7 @@ export default class SearchResultsWebPart extends BaseClientSideWebPart<ISearchR
             if (this.context.pageContext.legacyPageContext.userTimeZoneData) {
                 this._timeZoneBias.UserBias = this.context.pageContext.legacyPageContext.userTimeZoneData.Bias;
                 this._timeZoneBias.UserDST = this.context.pageContext.legacyPageContext.userTimeZoneData.DaylightBias;
-                this._timeZoneBias.Id = this.context.pageContext.legacyPageContext.webTimeZoneData.Id
+                this._timeZoneBias.Id = this.context.pageContext.legacyPageContext.webTimeZoneData.Id;
             }
 
             this._searchService = new SearchService(this.context.pageContext, this.context.spHttpClient);
@@ -1472,7 +1472,6 @@ export default class SearchResultsWebPart extends BaseClientSideWebPart<ISearchR
 
         const searchResultSourceData: ISearchResultSourceData = {
             queryKeywords: this.properties.queryKeywords.tryGetValue(),
-            queryModification: this._resultService.results.QueryModification,
             refinementResults: (this._resultService && this._resultService.results) ? this._resultService.results.RefinementResults : [],
             paginationInformation: (this._resultService && this._resultService.results) ? this._resultService.results.PaginationInformation : {
                 CurrentPage: 1,
